Move some more classes into apiimpl package.
