+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/apiimpl/grid/GridScrollModeUtil.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.platform.common.internal.grid;
+package com.refinedmods.refinedstorage2.platform.apiimpl.grid;
 
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
