+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/internal/grid/view/FabricFluidGridResourceFactory.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage2.platform.fabric.internal.grid.view;
 
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
-import com.refinedmods.refinedstorage2.platform.common.internal.grid.view.FluidGridResourceFactory;
+import com.refinedmods.refinedstorage2.platform.apiimpl.grid.view.FluidGridResourceFactory;
 
 import java.util.stream.Collectors;
