+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/internal/grid/view/FabricItemGridResourceFactory.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage2.platform.fabric.internal.grid.view;
 
-import com.refinedmods.refinedstorage2.platform.common.internal.grid.view.ItemGridResourceFactory;
+import com.refinedmods.refinedstorage2.platform.apiimpl.grid.view.ItemGridResourceFactory;
 
 import java.util.Optional;
