+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/internal/grid/view/ForgeItemGridResourceFactory.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage2.platform.forge.internal.grid.view;
 
-import com.refinedmods.refinedstorage2.platform.common.internal.grid.view.ItemGridResourceFactory;
+import com.refinedmods.refinedstorage2.platform.apiimpl.grid.view.ItemGridResourceFactory;
 
 import java.util.Optional;
